"kona_tmon.c: Fix build warnings"
[Defect / Enhancement / New Feature]

[Problem]
Warning messages about assignment between different
type variables in drivers/hwmon/kona_tmon.c

[Solution]
Removed unwanted casting

[Reviewers]
Kernel BSP team

[Testing]
    [Test Cases]
    1.No warning messages during build
    2.Flashed zImage on android_hawaii_garnet_edn010
      and basic working of device was normal

    [Label/CL# Synced To]

    [Platform]

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/37674
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ drivers/hwmon/kona_tmon.c
@@ -776,7 +776,7 @@ static int kona_tmon_probe(struct platform_device *pdev)
 			kfree(pdata);
 			goto err_free_dev_mem;
 		}
-		pdata->base_addr = (u32)ioremap(iomem->start,
+		pdata->base_addr = ioremap(iomem->start,
 					resource_size(iomem));
 		if (!pdata->base_addr) {
 			tmon_dbg(TMON_LOG_ERR, "unable to map in registers\n");
@@ -803,7 +803,7 @@ static int kona_tmon_probe(struct platform_device *pdev)
 			goto err_free_dev_mem;
 		}
 		val = *(addr + 1);
-		pdata->chipreg_addr = (u32)ioremap(be32_to_cpu(*addr),
+		pdata->chipreg_addr = ioremap(be32_to_cpu(*addr),
 				be32_to_cpu(val));
 
 		ret = of_property_read_u32(pdev->dev.of_node,
